requirements.txt
python-dotenv==0.19.0
Flask==2.0.1
requests==2.25.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.1
regex==2021.4.4
psycopg2==2.9.1